2.2.6 Injected Money into existing code

+++ src/main/java/io/legacyfighter/cabs/service/ClaimService.java
@@ -92,7 +92,7 @@ public class ClaimService {
     @Transactional
     public Claim tryToResolveAutomatically(Long id) {
         Claim claim = find(id);
-        if(claimRepository.findByOwnerAndTransit(claim.getOwner(), claim.getTransit()).size() > 1) {
+        if (claimRepository.findByOwnerAndTransit(claim.getOwner(), claim.getTransit()).size() > 1) {
             claim.setStatus(ESCALATED);
             claim.setCompletionDate(Instant.now());
             claim.setChangeDate(Instant.now());
@@ -108,7 +108,7 @@ public class ClaimService {
             return claim;
         }
         if (claim.getOwner().getType().equals(Client.Type.VIP)) {
-            if (claim.getTransit().getPrice() < appProperties.getAutomaticRefundForVipThreshold()) {
+            if (claim.getTransit().getPrice().toInt() < appProperties.getAutomaticRefundForVipThreshold()) {
                 claim.setStatus(REFUNDED);
                 claim.setCompletionDate(Instant.now());
                 claim.setChangeDate(Instant.now());
@@ -124,7 +124,7 @@ public class ClaimService {
             }
         } else {
             if (transitRepository.findByClient(claim.getOwner()).size() >= appProperties.getNoOfTransitsForClaimAutomaticRefund()) {
-                if (claim.getTransit().getPrice() < appProperties.getAutomaticRefundForVipThreshold()) {
+                if (claim.getTransit().getPrice().toInt() < appProperties.getAutomaticRefundForVipThreshold()) {
                     claim.setStatus(REFUNDED);
                     claim.setCompletionDate(Instant.now());
                     claim.setChangeDate(Instant.now());
